Fix tooltip not showing on first position
I have editor with 0 padding and absolutely positioned tooltips. The autocomplete tooltip was not showing because it's x position was the same as editor left. This should be strict inequality. I don't have a minimal repro but we could probably put it together easily. I'm not sure about the other inequalities and whether they should be fixed too.