Added an id to identify tooltips
Currently, the identity of the Tooltip.create function is used to
identify tooltips. When the function's identity changes between
view updates, this leads to creating new tooltips instead of
updating existing ones. This adds an `id` which is also used to identify
the known tooltip in addition to the create function